# api/openapi.yaml
openapi: 3.1.0
...
info:
  title: menu-planner API
  version: 1.0.0
  description: |
    献立提案アプリのREST API。

    認証は HttpOnly Cookie（access_token / refresh_token）で行う。
    エラーはすべて RFC 7807 の problem+json で返す。
servers:
  - url: /
    description: フロントと同一オリジン（Vite dev server が /api を backend に転送する）
...
paths:
...
  /api/v1/auth/signup:
    post:
      tags: [auth]
      summary: メールとパスワードで登録する
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialsRequest'
      responses:
        '201':
          description: 登録したユーザー。認証Cookieを設定する。
          headers:
            Set-Cookie:
              $ref: '#/components/headers/SetAuthCookies'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: メールアドレスが登録済み
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: メールとパスワードでログインする
      description: |
        存在しないメールもパスワード違いも同じ 401 に丸める。
        どちらであるかを返すと、登録済みメールを推測できてしまうため。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialsRequest'
      responses:
        '200':
          description: ログインしたユーザー。認証Cookieを設定する。
          headers:
            Set-Cookie:
              $ref: '#/components/headers/SetAuthCookies'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      summary: リフレッシュトークンでアクセストークンを再発行する
      description: refresh_token Cookie を使う。このCookieは /api/v1/auth にのみ送られる。
      security: []
      responses:
        '204':
          description: 新しいアクセストークンをCookieに設定した
          headers:
            Set-Cookie:
              $ref: '#/components/headers/SetAuthCookies'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/auth/google/callback:
    get:
      tags: [auth]
      summary: Google認証のコールバック
      description: |
        state を定数時間で照合し、コードをトークンに交換する。
        メール未確認のGoogleアカウントは受け付けない（他人のメールでの乗っ取りを防ぐ）。
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        '302':
          description: フロントエンドへのリダイレクト。認証Cookieを設定する。
          headers:
            Location:
              schema:
                type: string
                format: uri
            Set-Cookie:
              $ref: '#/components/headers/SetAuthCookies'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/me:
    get:
      tags: [auth]
      summary: 現在のユーザーを取得する
      responses:
        '200':
          description: 現在のユーザー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    accessCookie:
      type: apiKey
      in: cookie
      name: access_token
      description: ログイン時に設定される HttpOnly Cookie。
...
  headers:
    SetAuthCookies:
      description: |
        access_token（15分・path=/）と refresh_token（30日・path=/api/v1/auth）。
        いずれも HttpOnly / Secure / SameSite=Lax。
      schema:
        type: string
  responses:
    BadRequest:
      description: リクエストが不正
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: 認証が必要、または認証に失敗した
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    CredentialsRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          description: 8文字以上・72バイト以下。
          minLength: 8
    User:
      type: object
      required: [id, email, displayName, plan]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        plan:
          type: string
          enum: [free, premium]
          description: |
            契約プラン。上限そのものは返さない。フロントが件数を持つと
            サーバとの二重管理になるため、上限に達したことは 409 の本文で伝える。
    UserResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: '#/components/schemas/User'
...
    Problem:
      type: object
      description: RFC 7807 の Problem Details。
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: エラー種別を識別するURI。
        title:
          type: string
          description: 人間が読める短い要約。同じ type なら同じ title。
        status:
          type: integer
        detail:
          type: string
          description: この発生に固有の説明。

# 明示的に security: [] を書いたもの以外は認証必須。
security:
  - accessCookie: []